Fixed overcorrecting on low temps sometimes
Basically, there are times when your model will be very certain, but not be stuck in a loop, such as when reciting digits of pi or reciting other facts. So we instituted a range in which the "shock" can occur and aligned it with the range where loops usually occur.